Fix#2.4.1/chinh sua restaurant

+++ database/migrations/2018_08_18_012336_create_restaurants_table.php
@@ -33,7 +33,7 @@ class CreateRestaurantsTable extends Migration
             $table->time('closeTime');
             $table->string('shortDescription');
             $table->text('description');
-            $table->bigInteger('numberTable');
+            $table->int('numberTable');
             $table->float('lat');
             $table->float('lng');
             $table->integer('status')->default(1); // 1 hoạt động, 0 đã xóa, 2 ngừng bán
